+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-24 15:27:48
-// Commit: perf(core): apply RepaintBoundary on animated core widgets
+// Updated at: 2026-08-24 16:12:22
+// Commit: fix(core): prevent double tap duplicate action in core
 
 class coreComponentState {
-  final String commitRef = '3893';
-  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
-  final DateTime timestamp = DateTime.parse('2026-08-24T15:27:48+05:00');
+  final String commitRef = '3897';
+  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
+  final DateTime timestamp = DateTime.parse('2026-08-24T16:12:22+05:00');
 }